+++ src/Trail.java
@@ -13,7 +13,7 @@ public class Trail extends GameObject{
 	int height;
 	float life;
 	
-	public Trail(int x,int y,ID id,Color color,int width,int height,float life,Handler handler){
+	public Trail(float x,float y,ID id,Color color,int width,int height,float life,Handler handler){
 		super(x,y,id);
 		this.color = color;
 		this.handler = handler;
@@ -35,7 +35,7 @@ public class Trail extends GameObject{
 		Graphics2D g2d = (Graphics2D) g;
 		g2d.setComposite(makeTransparent(alpha));
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x, (int)y, width, height);
 		
 	}
